refactor to be able to easily add more projects
